자기 게시물 수정 및 삭제

+++ src/main/java/com/example/prg1/board/serivce/BoardService.java
@@ -85,18 +85,37 @@ public class BoardService {
         return dto;
     }
 
-    public void remove(Integer id) {
-        boardRepository.deleteById(id);
+    public boolean remove(Integer id, MemberDto user) {
+        if (user != null) {
+            Member db = boardRepository.findById(id).get().getWriter();
+
+            if (db.getId().equals(user.getId())) {
+                boardRepository.deleteById(id);
+                return true;
+            }
+
+        }
+
+        return false;
     }
 
-    public void update(BoardForm data) {
-        // 조회
-        Board board = boardRepository.findById(data.getId()).get();
-        // 수정
-        board.setTitle(data.getTitle());
-        board.setContent(data.getContent());
+    public boolean update(BoardForm data, MemberDto user) {
 
-        // 저장
-        boardRepository.save(board);
+        if (user != null){
+
+            // 조회
+            Board board = boardRepository.findById(data.getId()).get();
+            if (board.getWriter().getId().equals(user.getId())) {
+    //            System.out.println("11111111111111111111111111111111111");
+                // 수정
+                board.setTitle(data.getTitle());
+                board.setContent(data.getContent());
+                // 저장
+                boardRepository.save(board);
+                return true;
+            }
+
+        }
+        return false;
     }
 }
